// frontend/generated/models/ProjectCreateDto.ts
export interface ProjectCreateDto {
    /**
     * 
     * @type {number}
     * @memberof ProjectCreateDto
     */
    projectCategoryId?: number;
    /**
     * 
     * @type {string}
     * @memberof ProjectCreateDto
     */
    title?: string;
    /**
     * 
     * @type {string}
     * @memberof ProjectCreateDto
     */
    description?: string;
    /**
     * 
     * @type {Date}
     * @memberof ProjectCreateDto
     */
    startDate?: Date;
    /**
     * 
     * @type {number}
     * @memberof ProjectCreateDto
     */
    minBudget?: number;
    /**
     * 
     * @type {number}
     * @memberof ProjectCreateDto
     */
    maxBudget?: number;
    /**
     * 
     * @type {Array<SkillSetDto>}
     * @memberof ProjectCreateDto
     */
    requiredSkills?: Array<SkillSetDto>;
    /**
     * 
     * @type {Array<MilestoneCreateDto>}
     * @memberof ProjectCreateDto
     */
    milestones?: Array<MilestoneCreateDto>;
}