// frontend/generated/models/SkillSetDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Freelance Platform API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { ProficiencyEnum } from './ProficiencyEnum';
import {
    ProficiencyEnumFromJSON,
    ProficiencyEnumFromJSONTyped,
    ProficiencyEnumToJSON,
    ProficiencyEnumToJSONTyped,
} from './ProficiencyEnum';

/**
 * 
 * @export
 * @interface SkillSetDto
 */
export interface SkillSetDto {
    /**
     * 
     * @type {number}
     * @memberof SkillSetDto
     */
    skillId?: number;
    /**
     * 
     * @type {ProficiencyEnum}
     * @memberof SkillSetDto
     */
    proficiency?: ProficiencyEnum;
}



/**
 * Check if a given object implements the SkillSetDto interface.
 */
export function instanceOfSkillSetDto(value: object): value is SkillSetDto {
    return true;
}

export function SkillSetDtoFromJSON(json: any): SkillSetDto {
    return SkillSetDtoFromJSONTyped(json, false);
}

export function SkillSetDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): SkillSetDto {
    if (json == null) {
        return json;
    }
    return {
        
        'skillId': json['skillId'] == null ? undefined : json['skillId'],
        'proficiency': json['proficiency'] == null ? undefined : ProficiencyEnumFromJSON(json['proficiency']),
    };
}

export function SkillSetDtoToJSON(json: any): SkillSetDto {
    return SkillSetDtoToJSONTyped(json, false);
}

export function SkillSetDtoToJSONTyped(value?: SkillSetDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'skillId': value['skillId'],
        'proficiency': ProficiencyEnumToJSON(value['proficiency']),
    };
}