// frontend/generated/models/MilestoneCreateDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Freelance Platform API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface MilestoneCreateDto
 */
export interface MilestoneCreateDto {
    /**
     * 
     * @type {string}
     * @memberof MilestoneCreateDto
     */
    description?: string;
    /**
     * 
     * @type {string}
     * @memberof MilestoneCreateDto
     */
    title?: string;
    /**
     * 
     * @type {number}
     * @memberof MilestoneCreateDto
     */
    budgetRatio?: number;
    /**
     * 
     * @type {Date}
     * @memberof MilestoneCreateDto
     */
    deadline?: Date;
}

/**
 * Check if a given object implements the MilestoneCreateDto interface.
 */
export function instanceOfMilestoneCreateDto(value: object): value is MilestoneCreateDto {
    return true;
}

export function MilestoneCreateDtoFromJSON(json: any): MilestoneCreateDto {
    return MilestoneCreateDtoFromJSONTyped(json, false);
}

export function MilestoneCreateDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): MilestoneCreateDto {
    if (json == null) {
        return json;
    }
    return {
        
        'description': json['description'] == null ? undefined : json['description'],
        'title': json['title'] == null ? undefined : json['title'],
        'budgetRatio': json['budgetRatio'] == null ? undefined : json['budgetRatio'],
        'deadline': json['deadline'] == null ? undefined : (new Date(json['deadline'])),
    };
}

export function MilestoneCreateDtoToJSON(json: any): MilestoneCreateDto {
    return MilestoneCreateDtoToJSONTyped(json, false);
}

export function MilestoneCreateDtoToJSONTyped(value?: MilestoneCreateDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'description': value['description'],
        'title': value['title'],
        'budgetRatio': value['budgetRatio'],
        'deadline': value['deadline'] == null ? undefined : ((value['deadline']).toISOString()),
    };
}