// frontend/generated/models/ProficiencyEnum.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Freelance Platform API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const ProficiencyEnum = {
    Beginner: 'BEGINNER',
    Intermediate: 'INTERMEDIATE',
    Advanced: 'ADVANCED',
    Expert: 'EXPERT'
} as const;
export type ProficiencyEnum = typeof ProficiencyEnum[keyof typeof ProficiencyEnum];


export function instanceOfProficiencyEnum(value: any): boolean {
    for (const key in ProficiencyEnum) {
        if (Object.prototype.hasOwnProperty.call(ProficiencyEnum, key)) {
            if (ProficiencyEnum[key as keyof typeof ProficiencyEnum] === value) {
                return true;
            }
        }
    }
    return false;
}

export function ProficiencyEnumFromJSON(json: any): ProficiencyEnum {
    return ProficiencyEnumFromJSONTyped(json, false);
}

export function ProficiencyEnumFromJSONTyped(json: any, ignoreDiscriminator: boolean): ProficiencyEnum {
    return json as ProficiencyEnum;
}

export function ProficiencyEnumToJSON(value?: ProficiencyEnum | null): any {
    return value as any;
}

export function ProficiencyEnumToJSONTyped(value: any, ignoreDiscriminator: boolean): ProficiencyEnum {
    return value as ProficiencyEnum;
}